move util helpers into spec-helpers
