+++ server/src/model/bundle.spec.ts
@@ -1,6 +1,6 @@
 import expect from 'expect'
 import { Bundle, BundleValidationKind } from './bundle'
-import { expectErrors, first, loadSuccess, makeBundle, read } from './util.spec'
+import { expectErrors, first, loadSuccess, makeBundle, read } from './spec-helpers'
 
 interface BundleMakerInfo {
   version?: number
